Start using travis.ci to check auto generated docs
For starters only the auto generated part of the docs is regenerated in order
to check that the last author did regenerate and commit them.

Some dependencies of alot are mocked to speed up installation.  Especially
notmuch would need to be build manually as the version available on travis is
to old.