Set up several jobs to build docs and run tests in parallel
Two jobs are set up to run on travis.  Both run on Ubuntu trusty (14.04) and
python2.7.  One checks and builds the docs and the other is set up to run the
test suite (currently only a small config file is used to start and quit alot
again, much like `vim -c quit`).

The longer and more complicated shell scripts are still kept in the travis
file (with yaml multi line strings).  They get the shell option `set -e` to
get the "short circuit" behaviour otherwise present with single statements
(also known from make files) to fail the whole job as soon as any single line
fails.

Currently the tests are run in a full VM as we need sudo to manually build a
new enough version of notmuch.  This can be changed when this PR is resolved:
https://github.com/travis-ci/apt-package-whitelist/issues/3895