use open() as context manager
This uses open() as a context manager instead of calling close()
explicitly. This has the advantage of closing even in the event of an
exception, being easier to visually inspect for correctness, and being
the more modern idiom.

This does leave one case of file.close(), where FILE is opened in an if
tree.