commands/utils: Refactor exception case
This removes the use of range (originally I simply replaced it with
xrange, but the realized that the use of range was itself strange), and
replaces it with generators and iterators to create the dictionary. This
has the advantage of only creating one concrete instance (the choices
dict), and being slightly easier to read and understand, as well as not
needing to call len() repeatedly.