Move default notmuch config path to argparse parser setup
This is only possible for the notmuch config file and not for the alot config
file as the former is strictly required but the latter is not.  If the latter
does not exist it will be detected by the argument parser.  If the former is
not given on the command line the parser will return `None` and we can
manually check if the default config file exists and else just skip the alot
config file.