commands/global: Don't double prompt to close last buffer
Currently if global.quit_on_last_bclose is set to True (not the default),
and the last buffer is an envelope, then when 'd' or ':bclose' is
called, the user will be prompted twice to 'close without sending?'.

This patch fixes that by skipping the prompt in BufferCloseCommand, if
there is 1 buffer, and global.quit_on_last_bclose is True. It does this
by adding a private keyword to the ExitCommand() constructor, telling
that classes apply method to not prompt.

Fixes #979