tests: Instruct pylint to ignore a large swath of warnings
There are a number of things pylint warns on that absolutely make sense
to fix in production code, but for unittests they either don't matter
(like naming variables "foo"), can't be fixed (TestCase methods that
don't use self because they use a mock assert), or the descriptive names
violate PEP8. These are annoying and create noise, so tell pylint to
ignore them.