settings: Allow loading themes from system locations
This adds additional system locations to search for theme files, which
are either defined as the environment variable XDG_DATA_DIRS, or the
generic fallback (per the xdg base directories spec) as
/usr/local/share and /usr/share. We then add alot/themees to those paths
to search for themes in.  This allows a package manager to install the
theme files system wide along with alot, and for alot to load them from
the system wide directories.

The Archlinux alot-git package is already installing the packages like
this.