tests/crypto: Fix GetKey.test_missing_key
'z', the value currently passed to get_key isn't a valid value, so it
raises GPGMEError with a code of INV_VALUE. However, if an actual email
is passed a KeyNotFound exception is raised instead.

The existing test is valid and should remain, since it catches a
potential bug, but it doesn't test for a missing key, it tests for an
invalid key. This patch renames the existing test and adds a new test to
cover an actual missing key.