crypto: Don't sign when encrypting
This is a missed difference between pygpgme and gpg, gpg signs by
default when encrypting with an attached signature. Since this isn't
being toggled it's bad, it's a kind of information leak, and increases
the size of the mail for no reason.

NOTE: This is the kind of signature proposed by RFC 2440, which we're
missing tests for.