provide retagprompt command in thread mode
retagprompt in search mode allows to retag a whole thread. The new
thread mode version allows to do the same for each selected message.

The common command infers its mode from the UI object that it is passed
when applied. Therefore, the docstring (and everything else) is
mode-agnostic.

When this command command is registered for the wrong mode and applied
it raises a KeyError (on purpose).