update theme spec, default theme and tests
This patchseries adds a new buffer type for displaying notmuch's saved
querystrings. This commit adds the respective theming attributes used in
the new buffer. They are just copies of the taglist buffer really.

THIS BREAKS COMPATIBILITY WITH USER THEMES!
The theme files need to be updated to also have these new attributes in
order to conform with the updated spec.