commands/thread: Don't refresh after moving
For large threads (I can reproduce on threads with ~50 messages),
changing messages can become quite slow, especially if there are a
significant number of folded messages. Removing the call to refresh the
body after moving seems to fix this.

There may be side effects of this change, and it may be required in some
cases, just not all of them.